docs: move #slider over to the new page structure
