+++ packages/dnb-design-system-portal/src/pages/uilib/components/slider.md
@@ -1,15 +1,26 @@
 ---
 title: 'Slider'
 draft: false
-status: 'imp'
+# status: 'imp'
 order: 13
 ---
 
-<!--
-  ATTENTION: This file is auto generated by using "makeDemosFactory".
-  Do not change the content!
--->
+import Tabs from 'Tags/Tabs'
 
-import SliderPage from 'Src/uilib/components/demos/Slider'
+import SliderInfo from 'Pages/uilib/components/slider/slider-info'
+import SliderProperties from 'Pages/uilib/components/slider/slider-properties'
+import SliderEvents from 'Pages/uilib/components/slider/slider-events'
 
-<SliderPage />
+# Slider
+
+<Tabs>
+  <Tabs.Content>
+    <SliderInfo />
+  </Tabs.Content>
+  <Tabs.Content>
+    <SliderProperties />
+  </Tabs.Content>
+  <Tabs.Content>
+    <SliderEvents />
+  </Tabs.Content>
+</Tabs>
