+++ packages/dnb-design-system-portal/src/pages/uilib/components/slider/slider-events.md
@@ -1,0 +1,13 @@
+---
+draft: true
+---
+
+## Events
+
+| Events            | Description                                                                                                |
+| ----------------- | ---------------------------------------------------------------------------------------------------------- |
+| `on_init`         | _(optional)_ will be called once the component is ready to use. The callback value is a number `{ value }` |
+| `on_change`       | _(optional)_ will be called on state changes made by the user. The callback value is a number `{ value }`  |
+| `on_state_update` | _(optional)_ will be called once the parameter `value` changes its value.                                  |
+| `on_drag_start`   | _(optional)_ will be called once the user stops dragging.                                                  |
+| `on_drag_end`     | _(optional)_ will be called once the user starts dragging.                                                 |
