+++ packages/dnb-design-system-portal/src/pages/uilib/components/slider/slider-info.md
@@ -1,0 +1,13 @@
+---
+draft: true
+---
+
+import Examples from 'Pages/uilib/components/slider/Examples'
+
+## Description
+
+The `Slider` component provides a visual indication of adjustable value. A value can be adjusted (increased or decreased) by moving the drag handle along a track (usually horizontal or vertical). Remember to inform users that they can also adjust the value directly in the value input field (if it exists).
+
+## Demos
+
+<Examples />
