--- packages/dnb-design-system-portal/src/uilib/components/examples/Slider.txt
@@ -1,8 +1,0 @@
-<Slider
-  min={0}
-  max={100}
-  value={70}
-  step={10}
-  on_change={this.onChangeHandler}
-/>
-<Slider value="70" max="100" disabled />
