--- packages/dnb-ui-lib/src/components/slider/description.md
@@ -1,7 +1,0 @@
----
-status: 'wip'
----
-
-## Description
-
-Sliders provide a visual indication of adjustable content. A value can be adjusted (increased or decreased) by moving the drag handle along a track (usually horizontal or vertical). Remember to inform users that they can also adjust the value directly in the value input field (if it exists).
